リザルト！！！！！！！！完成！！！！！！！！！！！

+++ 00_project/source/retentionManager.cpp
@@ -104,7 +104,7 @@ void CRetentionManager::SetResult(const long nTime)
 //============================================================
 //	リザルト情報の設定処理
 //============================================================
-void CRetentionManager::SetScore(const long Score)
+void CRetentionManager::SetScore(const int Score)
 {
 	// 引数の経過タイムを設定
 	m_result.nScore = Score;
